separate repeatable from embedded hapi tests
